Rename hint content to hint
